fix(explore): round editable title width measurement
